src/ai/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pydantic-settings==2.*
python-multipart==0.0.*
httpx==0.28.*
Pillow==11.*
supabase==2.*
fal-client==0.5.*
python-dotenv==1.*
sentry-sdk[fastapi]==2.*
